+++ src/anemoi/graphs/nodes/builders/from_refined_icosahedron.py
@@ -12,6 +12,7 @@ from torch_geometric.data import HeteroData
 
 from anemoi.graphs.generate.hex_icosahedron import create_hex_nodes
 from anemoi.graphs.generate.masks import KNNAreaMaskBuilder
+from anemoi.graphs.generate.tri_icosahedron import create_stretched_tri_nodes
 from anemoi.graphs.generate.tri_icosahedron import create_tri_nodes
 from anemoi.graphs.nodes.builders.base import BaseNodeBuilder
 
@@ -25,8 +26,6 @@ class IcosahedralNodes(BaseNodeBuilder, ABC):
     ----------
     resolution : list[int] | int
         Refinement level of the mesh.
-    name : str
-        Name of the nodes.
     """
 
     def __init__(
@@ -53,13 +52,13 @@ class IcosahedralNodes(BaseNodeBuilder, ABC):
         return torch.tensor(coords_rad[self.node_ordering], dtype=torch.float32)
 
     @abstractmethod
-    def create_nodes(self) -> tuple[nx.DiGraph, np.ndarray, list[int]]: ...
+    def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]: ...
 
     def register_attributes(self, graph: HeteroData, config: DotDict) -> HeteroData:
         graph[self.name]["_resolutions"] = self.resolutions
         graph[self.name]["_nx_graph"] = self.nx_graph
         graph[self.name]["_node_ordering"] = self.node_ordering
-        graph[self.name]["_aoi_mask_builder"] = self.aoi_mask_builder
+        graph[self.name]["_area_mask_builder"] = self.area_mask_builder
         return super().register_attributes(graph, config)
 
 
@@ -68,7 +67,7 @@ class LimitedAreaIcosahedralNodes(IcosahedralNodes):
 
     Attributes
     ----------
-    aoi_mask_builder : KNNAreaMaskBuilder
+    area_mask_builder : KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
@@ -83,10 +82,10 @@ class LimitedAreaIcosahedralNodes(IcosahedralNodes):
 
         super().__init__(resolution, name)
 
-        self.aoi_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
+        self.area_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
 
     def register_nodes(self, graph: HeteroData) -> None:
-        self.aoi_mask_builder.fit(graph)
+        self.area_mask_builder.fit(graph)
         return super().register_nodes(graph)
 
 
@@ -117,12 +116,12 @@ class LimitedAreaTriNodes(LimitedAreaIcosahedralNodes):
 
     Parameters
     ----------
-    aoi_mask_builder: KNNAreaMaskBuilder
+    area_mask_builder: KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
     def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
-        return create_tri_nodes(resolution=max(self.resolutions), aoi_mask_builder=self.aoi_mask_builder)
+        return create_tri_nodes(resolution=max(self.resolutions), area_mask_builder=self.area_mask_builder)
 
 
 class LimitedAreaHexNodes(LimitedAreaIcosahedralNodes):
@@ -132,9 +131,54 @@ class LimitedAreaHexNodes(LimitedAreaIcosahedralNodes):
 
     Parameters
     ----------
-    aoi_mask_builder: KNNAreaMaskBuilder
+    area_mask_builder: KNNAreaMaskBuilder
         The area of interest mask builder.
     """
 
     def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
-        return create_hex_nodes(resolution=max(self.resolutions), aoi_mask_builder=self.aoi_mask_builder)
+        return create_hex_nodes(resolution=max(self.resolutions), area_mask_builder=self.area_mask_builder)
+
+
+class StretchedIcosahedronNodes(IcosahedralNodes):
+    """Nodes based on iterative refinements of an icosahedron with 2
+    different resolutions.
+
+    Attributes
+    ----------
+    area_mask_builder : KNNAreaMaskBuilder
+        The area of interest mask builder.
+    """
+
+    def __init__(
+        self,
+        global_resolution: int,
+        lam_resolution: int,
+        name: str,
+        reference_node_name: str,
+        mask_attr_name: str,
+        margin_radius_km: float = 100.0,
+    ) -> None:
+
+        super().__init__(lam_resolution, name)
+        self.global_resolution = global_resolution
+
+        self.area_mask_builder = KNNAreaMaskBuilder(reference_node_name, margin_radius_km, mask_attr_name)
+
+    def register_nodes(self, graph: HeteroData) -> None:
+        self.area_mask_builder.fit(graph)
+        return super().register_nodes(graph)
+
+
+class StretchedTriNodes(StretchedIcosahedronNodes):
+    """Nodes based on iterative refinements of an icosahedron with 2
+    different resolutions.
+
+    It depends on the trimesh Python library.
+    """
+
+    def create_nodes(self) -> tuple[nx.Graph, np.ndarray, list[int]]:
+        return create_stretched_tri_nodes(
+            base_resolution=self.global_resolution,
+            lam_resolution=max(self.resolutions),
+            area_mask_builder=self.area_mask_builder,
+        )